move: hooks -> into hooks folder, minor guestData refactor
