+++ src/components/Dashboard/PackCategory/TableCells/ItemNameCell/ItemNameCell.tsx
@@ -2,7 +2,7 @@ import { Table, Input, Popup, Icon, Button } from 'semantic-ui-react';
 import { useState } from 'react';
 import './ItemNameCell.css';
 import { GripButton } from '../../TableButtons/TableButtons';
-import { useUserContext } from '../../../../../views/Dashboard/useUserContext';
+import { useUserContext } from '../../../../../views/Dashboard/hooks/useUserContext';
 import Link from '../../../../../shared/ui/Link';
 import { InputEvent, SelectEvent } from '../../../../../shared/formHelpers';
